// src/app/models/users/Person.ts
import { Address } from './Address';
import { Phone } from './Phones';

export interface Person {
    id?: number;
    name?: string;
    secondName?: string;
    lastName?: string;
    secondLastName?: string;
    birthDate?: Date;
    status?: boolean;
    maritalStatus?: string;
    gender?: string;
    imagen?: string;
    dateToShow?: string;
    address?: Address[];
    phones?: Phone[];
}